trainer/pyproject.toml
[tool.poetry]
name = "gabate_brain"
version = "0.1.0"
description = "The brains behind GABATE"
authors = ["h3nnn4n <uber.renan@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^3.5.3"
python-dotenv = "^0.19.1"
dramatiq = {extras = ["redis", "watch"], version = "^1.15.0"}
periodiq = "^0.12.1"
python-decouple = "^3.8"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
mock = "^4.0.3"
celery = {extras = ["pytest"], version = "^5.1.2"}
black = "^21.9b0"
flake8 = "^4.0.1"
isort = "^5.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 121
